Extended the id range for npcs, now [400,700) will also be treated as NPC objects (see topic:170845 and bugreport:727). Silenced a compilation warning.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@12025 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/map.c
@@ -2338,7 +2338,7 @@ int map_readfromcache(struct map_data *m, FILE *fp)
 	if( i < header.map_count )
 	{
 		unsigned char *buf, *buf2;
-		unsigned int size, xy;
+		unsigned long size, xy;
 
 		m->xs = info.xs;
 		m->ys = info.ys;
